Split index.html into parts.  Slight mod to FPparse.py
-FPparse.py:
changed csv open format to string to 'rU' (something about the line
breaks in the file was causing an error when I run the script on my
mac).
Slightly changed the html output to generate an ordered list (instead
of paragraph elements starting with a number).  this is more semantic
and facilitates styling the bibliography.

-FPs.csv: added a few FPs and more DOIs

-index.html: removed css and js to external files for better
organization.

-created css and js folders with corresponding files.