Dynamic Sliders, simpler code
Sliders are now dynamically generated based on the "filters" object.
This allowed for much simpler code both in the initialization, as well
as in the slider event and the data filtering step. (plus, adding new
filter sliders is as easy as adding them to the filters object)

removed "styles strings" objects for simpler conditionals… if you end
up needing more than two style types for the tables, feel free to add
them back.

NOTE: the colorful spectrum graphic on the ex/em sliders needs
positioning work.  It only works in firefox right now but is out of
position in webkit browsers.